test(ui): move some jest configs in share
